共通用の型定義ファイルを作成
