+++ src/app/(main)/_types/index.ts
@@ -4,27 +4,3 @@ export interface FieldBaseProps {
   label: string;
   error?: FieldError;
 }
-
-export interface Post {
-  id: number;
-  title: string;
-  content: string;
-  thumbnailUrl: string;
-  created_at: string;
-  updated_at: string;
-  PostCategory: PostCategory[];
-}
-
-export interface PostCategory {
-  id: number;
-  post_id: number;
-  category_id: number;
-  createdAt: string;
-  updatedAt: string;
-  category: Category;
-}
-
-export interface Category {
-  id: number;
-  name: string;
-}
